requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.1
openpyxl==3.1.2
Werkzeug==3.0.1
gunicorn==21.2.0 